pyproject.toml
[tool.poetry]
name = "gaia-api"
version = "0.1.0"
description = "Community monitoring app for environmental issues"
authors = ["blackbear <zoeytu0919@gmail.com>"]
readme = "README.md"
packages = [{include = "gaia_api"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.2"
uvicorn = "^0.23.2"
sqlalchemy = "^2.0.21"
mysql-connector-python = "^8.1.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.9"
black = "^23.9.1"
flake8 = "^6.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
